A semester project for TMB102-DATA BASES
A 2-part project in which in the first part we had to design a data base in sql  for an airline based on the given relations that were given. Also the data base mush have adhered to the BCNF (Boyce-Codd normal form) principles.
The second part was implementing specific sql queries that were asked off us.