End semester RTS game in Unity for TMD133
Every action was scripted by me , although it is barebone and unfinished it taught me a decent amount  in how to develop strategy games, especially ones with  building aspects. In this file are the scripts that were used  as well some scriptable objects to more or less give a better picture.